+++ src/main/java/gift/product/presentation/WishListController.java
@@ -47,7 +47,7 @@ public class WishListController {
 
     @JwtAuthenticated
     @PostMapping("/{wishListId}/add/{productId}/{optionId}/{quentity}")
-    public ResponseEntity<?> addProductToWishList(@PathVariable Long wishListId, @PathVariable Long productId, @PathVariable Long optionId, @PathVariable int quentity) {
+    public ResponseEntity<?> addProductToWishList(@PathVariable Long wishListId, @PathVariable Long productId, @PathVariable Long optionId, @PathVariable Long quentity) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         Long userId = Long.valueOf(authentication.getName());
         wishListService.addProductToWishList(userId, wishListId, productId, optionId, quentity);
